setup file structure for gemfile
* also reenabled command line call ruby compare.rb gitname,gitname...
