+++ lib/comparer.rb
@@ -1,0 +1,4 @@
+require_relative "user"
+
+module Comparer
+end
